refactor[#186] : LogoutButton features -> entities 레이어 이동
