--- src/features/profile/index.ts
@@ -1,1 +1,0 @@
-export * from "./LogoutButton";
